requirements.txt
accelerate>=0.16.0
torchvision~=0.17.1
transformers>=4.25.1
datasets~=2.18.0
ftfy
tensorboard
Jinja2
peft==0.7.0
numpy~=1.26.4
torch~=2.2.1
# Not required as we are installing directly from git source.
# diffusers~=0.27.2
packaging~=24.0
tqdm~=4.66.2
librosa~=0.10.1
wandb
git+https://github.com/huggingface/diffusers
pillow==9.5.0